Argon2id support
This adds support for the argon2id algorithm, which seems to be the
[preferred algorithm][] for general use and password encryption.

There are a few differences between this implementation and the argon2i
implementation:

 * No maximum salt size.
 * Include parameters in output regardless of whether they match the
   defaults.

These match the output from the [winning implementation][] of the
Password Hashing Contest.

I, Daniel Parks, transfer all copyrights I have on this code to Philipp
Brüll.

[winning implementation]: https://github.com/P-H-C/phc-winner-argon2
[preferred algorithm]: https://tools.ietf.org/html/draft-irtf-cfrg-argon2-02#section-4